Move IssuesTable to its own file
